Crate wirefish

Packet sniffing application built with Tauri

Functionalities

List all available network interfaces

Select a network interface

Start the sniffing process

Stop the sniffing process

Pause the sniffing process

Resume the sniffing process

Generate a .csv report of the collected data

Errors These are the errors that can occur during the sniffing process, grouped by the action that can cause them:

Select interface

Inexistent

Start sniffing

Without prior selection of the interface

(?) Unhandled channel type

(?) Failed channel creation

Empty interface

Re-Start sniffing

Same interface

Another interface never selected

Another interface selected previously

Stop Sniffing

Sniffing process wasn’t started

Generate report

Generation failed (Permission denied)

Modules

filtering 🔒

Retrieve packets filtered by their attributes

report 🔒

This crate allows the generation of a periodic report in .csv format The report highlights the the first and last timestamp, the amount of traffic, and the protocols of data exchange for all connections identified by (Source IP, Destination IP, Source Port, Destination Port)

Structs

SniffingInfo 🔒

Informations about the selected network interface

SniffingState

Sniffing channel and data collected by the sniffing process

Enums

SniffingError

Errors that can occur during the sniffing process

Constants

CONFIG 🔒

Functions

generate_report 🔒

Produces or updates a .csv report with the data collected since the last report generation

get_interfaces_list 🔒

Returns the list of all available network interfaces

main 🔒

select_interface 🔒

Selection of a network interface among all the available ones

start_sniffing 🔒

Instantiates a new thread that will execute the sniffing process

stop_sniffing 🔒

Terminates (stop: true) or Pauses (stop: false) the sniffing process