Module wirefish::report::data

Data structures used to write a report

Structs

PacketExchange

Data structure describing the list of protocols, total bytes, and timestamps of the first and last packet exchange in a connection

SourceDestination

Ip addresses and port numbers of source and destination of a packet exchange