Module wirefish::report

This crate allows the generation of a periodic report in .csv format The report highlights the the first and last timestamp, the amount of traffic, and the protocols of data exchange for all connections identified by (Source IP, Destination IP, Source Port, Destination Port)

The following example describes how to use the defined data structures and generate a report:

use report::{
   data::{PacketExchange, SourceDestination},
   write_report
};
use chrono::Local;

fn main() {

    // Create a hashmap to hold all packet exchanges
    let exchanged_packets = HashMap::<SourceDestination, PacketExchange>::new();

    // Create Sender-Receiver pair
    let source_destination = SourceDestination::new(
        network_source,
        network_destination,
        transport_source,
        transport_destination
    );

    // Insert a packet in the hashmap
    let now = Local::now();
    let transmitted_bytes = 1024;
    let mut protocol = vec![String::from("IPv4"), String::from("UDP")];
    exchanged_packets
        .entry(source_destination)
        .and_modify(|exchange| {
            // Update data about exchanged packets
            exchange.add_packet(protocols.clone(), transmitted_bytes, now)
        })
        // Create a new Packet Exchange if none was found
        .or_insert(PacketExchange::new(protocols, transmitted_bytes, now));

    // Generate report
    let report_path = "./path/to/report.csv";
    let mut first_generation = true; // Only the first time, this adds the csv header
    write_report(report_path, exchanged_packets, first_generation);

    // From the second time onwards
    first_generation = false;

    // .. Add packets exchange ..
    write_report(report_path, exchanged_packets, first_generation);

    // .. Add packets exchange ..
    write_report(report_path, exchanged_packets, first_generation);
}

Modules

data

Data structures used to write a report

Functions

get_sender_receiver

Returns (Source IP, Destination IP, Source Port, Destination Port, and Protocols) contained in a packet

write_report

Appends data to a report file, creates the file if it doesn’t exist